Port Lekmod game core to native macOS and fix gameplay ABI compatibility

+++ LEKMOD_DLL/CvGameCoreDLL_Expansion2/CvGame.cpp
@@ -11147,7 +11147,7 @@ void CvGame::Write(FDataStream& kStream) const
 			CvAssertMsg(false,"Saved game database exists, but could not open it!");
 		}
 
-		long nilSize = 0;
+		int nilSize = 0;
 		kStream << nilSize;
 	}
